<feat #116> Answer, 리허설 재시도를 위한 sequence 필드 추가

+++ src/main/java/com/prography/minari/answer/entity/Answer.java
@@ -20,7 +20,7 @@ public class Answer extends BaseTimeEntity {
     private Long id;
 
     @Column(name = "sequence")
-    private Integer sequence;
+    private Long sequence;
 
     @Column(name = "reply", length = 1000)
     private String reply;
